senior_supplement/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
boto3==1.34.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
faiss-cpu==1.7.4
youtube-transcript-api==0.6.1
beautifulsoup4==4.12.2
lxml==4.9.3
psycopg2-binary==2.9.9